feat: add use cleanup_message, rewrite_message; refactoring logger

+++ src/plugins/user/sources_monitoring/edited_message/media_group.py
@@ -3,12 +3,12 @@ import logging
 from pyrogram import Client, filters
 from pyrogram.types import Message
 
-from common import get_shortened_text
-from models import CategoryMessageHistory
+from models import CategoryMessageHistory, Source
 from plugins.user.sources_monitoring.edited_message.common import (
     get_history_obj,
     is_out_edit_timeout,
     logging_on_startup,
+    set_edited_on_history,
     try_set_blocked,
 )
 from plugins.user.sources_monitoring.new_message.media_group import (
@@ -16,6 +16,8 @@ from plugins.user.sources_monitoring.new_message.media_group import (
 )
 from plugins.user.utils import custom_filters
 from plugins.user.utils.blocking import blocking_received_media_groups
+from plugins.user.utils.cleanup import cleanup_message
+from plugins.user.utils.rewriter import rewrite_message
 
 
 @Client.on_edited_message(
@@ -35,13 +37,24 @@ async def edited_media_group_message(client: Client, message: Message):
     if not blocked:
         return
 
-    messages_to_delete = get_message_to_delete(history_obj)
+    if history_obj.rewritten:
+        if message.caption:
+            source = Source.get(tg_id=message.chat.id)
+            cleanup_message(message, source)
+            rewrite_message(message)
 
-    await client.delete_messages(history_obj.category.tg_id, messages_to_delete)
-
-    set_deleted_on_history(history_obj, messages_to_delete)
-
-    await send_message_to_category(client, message)
+        await client.edit_message_caption(
+            chat_id=history_obj.category.tg_id,
+            message_id=history_obj.message_id,
+            caption=message.caption,
+            caption_entities=message.caption_entities,
+        )
+        set_edited_on_history(history_obj)
+    else:
+        messages_to_delete = get_message_to_delete(history_obj)
+        await client.delete_messages(history_obj.category.tg_id, messages_to_delete)
+        set_deleted_on_history(history_obj, messages_to_delete)
+        await send_message_to_category(client, message)
 
     blocked.remove(message.media_group_id or message.id)
 
@@ -70,23 +83,21 @@ def set_deleted_on_history(
     )
     query.execute()
     logging.info(
-        f'Сообщение {history_obj.source_message_id} из источника'
-        f' {get_shortened_text(history_obj.source.title, 20)} {history_obj.source.tg_id} было'
-        ' изменено. Все сообщения медиа-группы были удалены.'
-        f' {get_shortened_text(history_obj.category.title, 20)} {history_obj.category.tg_id}'
+        'Источник %s изменил сообщение %s. Все сообщения медиагруппы были удалены из категории %s',
+        history_obj.source_chat_id,
+        history_obj.source_message_id,
+        history_obj.category_id,
     )
 
 
 async def send_message_to_category(
     client: Client,
     message: Message,
-    # history_obj: CategoryMessageHistory,
 ) -> None:
     if b := blocking_received_media_groups.get(message.chat.id):
         b.remove(message.media_group_id)
     await new_media_group_message(
         client,
         message,
-        is_resending=True,  # Удалить
-        # history_obj=history_obj,  # Замена is_resending для возможности редактирования поста
+        is_resending=True,
     )
